feat(bootstrap-artifact): Install bootstrap Artifact on first start-up.
On start-up, Mender checks for the existence of an special bootstrap
Artifact in path `/var/lib/mender/bootstrap.mender` and installs it in
order to initialize the device database.

This applies both for `daemon` start and cli commands `bootstrap` and
`install`.

The Artifact is not installed if the device already has a database. When
the Artifact is not found (and the database is empty) the database is
initialized with `artifact-name=unknown`.

In addition, Mender can also understand other kinds of "empty" Artifacts
and install them either in managed or standalone modes.

Changelog: Commit
Ticket: MEN-2583

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>