fix: the periodic update check bug in checkWaitState
Simplified the `checkWait` logic a lot by moving the last attempt out of the
`state.go` file, and into `daemon.go`.

While maybe not completely kosher on the surface, but if you look through the
code, you will realize that this does indeed simplify our somewhat fragile logic
here a lot, and should now be fairly fool-proof, as opposed to previously, where
it has broken for us multiple times.

I also, added an exception for the manual forced update and inventory checks
from the cli, so that they do no longer interfere with the regularity of the
intervals at which the update checks are done by the state machine. I'm sure we
could do something clever here, but I figured a simple bool if, INF-1000 style
would suffice here.

Changelog: The client update and inventory checks are now unaffected by the use
of the `check-update` and `send-inventory` commands. While previously, this
would both move the intervals at which checks we're done, and also extend them
beyond the expected polling intervals configured.

Ticket: MEN-5547

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>