feat: Support multiple verification keys
https://hub.mender.io/t/multiple-artifactverifykeys/4309/6

Creates a new client config parameter called ArtifactVerifyKeys that is
a list of paths to keys. Any matching key can be used to verify an
artifact - e.g. if 5 verification keys are provided, only 1 needs to
match to verify the artifact.

This also adds the ability to filter keys by update type - so some keys
can be specified for use only for certain update types.

This enables a few key features:
- key rotation with a longer grace period by specifying multiple keys
  while rotating keys.
- separate verification keys for different update types.
- verifying artifacts signed for multiple environments/devices.

Changelog: Commit
Ticket: None

Signed-off-by: Michael Ho <callmemikeh@gmail.com>