fix: Artifact name reverting to "unknown" after upgrading from 1.7.x.
After upgrading from 1.7.x and booting the new operating system, the
client commits the artifact name into the database, but no Provides
values, since 1.7.x compatible artifacts do not have any. This leads
to a database which contains these values:

```
{
  "artifact_name": "new-name",
  "provides": null
}
```

Typically the null value will not occur when moving between later
versions of the client that have Provides values, but it's entirely
legal, and can be forced by stripping the artifact of all Provides
values (using compatibility options).

Since it is legal, it is no problem in itself, but reveals a really
sneaky bug in our JSON parsing: If, and only if, the value of the JSON
document is `null` (the special value, not the string), the JSON
parser replaces the map with a new empty map instead of leaving it
intact, like it does for all other JSON documents. This sort of makes
sense, but is undocumented, and a quite nasty corner case.

This leads the function to return what looks like a completely empty
database, even though in fact the artifact name is there. The code
therefore concludes the database is uninitalized and takes the path to
install the bootstrap artifact, which replaces the artifact name, even
though we already have one.

What makes this bug particularly nasty is that installing the
bootstrap artifact happens at launch time, which means the daemon
could be running for hours unaffected, and the bug would not be
triggered until you either restarted the client, or ran it on the
command line. And this is also why our tests did not initially catch
it. They started catching it after `show-artifact` was changed to
trigger the bootstrap artifact installation though.

Changelog: None
Ticket: QA-426

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>