feat: do not check if the artifact is already installed
The Mender Client checked if the artifact is already installed, comparing
the Artifact's name of the update with the one of the latest installed
on the device. However, since the introduction of Update Modules and
application updates, this check is weak because it relies on the latest
installed artifact only. We want to move all the `already-installed`
logic to the backend. Therefore, we remove this redundant check from
the client and rely on the Mender Server only.

Changelog: Title
Ticket: MEN-6129

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>