fix: Rework handling of platform-dependent implementations
The goal is to have the platform-dependent code isolated with the
internal interface, the 'common_' libs and headers, encapsulating
the implementation details.

That can be achieved by:

  - header files in the `common` directory declare the classes and
    functions for the common interfaces,

  - separate implementations provide definitions for the classes
    and functions declared in the common headers,

  - build system chooses the right implementation for each common
    interface and compiles the common interface library
    accordingly and

  - #ifdefs are used in the common interface header files if the
    individual implementations need specific **private** methods
    or attributes.

With the above, the code that uses a common interface only links
to the particular `common_IFACE` library and includes the
`common/IFACE.hpp` header.

The commit also renames the `common/something/impl` directories
to `common/something/platform` because having this subdirectory
actually means the code for `common/something` is
platform-dependent.

Ticket: MEN-6112
Changelog: None
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>