feat: templates and classes for error handling
Because of the target platforms, Mender client code cannot rely
on C++ exceptions. This means we need some other means of error
handling. There is the 'std::expected' templated class that can
wraps either an expected value or an error value, but it is only
in C++23 so not available everywhere the code base will be built.

This commit introduces a custom implementation of a similar
templated class called Expected together with a template for
errors consisting of an error code together with a string error
message.

Ticket: MEN-6112
Changelog: None
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>