chore: Add IO interfaces and convenience functions.
Since the interfaces need a non-template error type, we need to settle
on one and make some changes to our own error type. First of all, we
want to base it on the standard C++11 `error_condition` error type, so
that we can use error codes from there. This means we need to remove
the enum requirement from our error type.

Also make a few other convenience fixes such as the ability to get a
no-error object by default construction, a `StdError` typedef, and a
bool operator for easy testing in conditionals.

With that, implement the `Reader`, `Writer` and `ReadWriter`
interfaces, which are inspired from the Golang interfaces of the same
names, as well as the `Copy` functions, and add some some tests for
them.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>