chore: Switch Reader/Writer interfaces to use iterators.
This is more in line with how the std interfaces work, and saves us
from having to resize containers before and after passing them into
the Read and Write functions. The resulting code reduction can be
clearly seen.

The tests are a bit more complicated because they do deep comparisons,
and comparing ranges of iterators is harder than just two values. We
cannot just match arguments one by one anymore, but we need to use a
custom function where we call `std::equal`. This is limited to the
io_test though, and I don't think it affects production code.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>