feat(tar): Add the tar reader library
This adds a tar reader library implementation used to support reading Artifacts,
in the upcoming Artifact library.

The implementation is a standard mender interface surrounding libarchive.

The interface is dead-simple and looks like this:

```
Reader(io::Reader)
Reader.Next() -> TarEntry (implements Reader)
```

Where the entry holds the name of the entry, and the size information of the
embedded data.

Ticket: MEN-6200
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>