chore: Fix incorrect filling of buffer.
Turns out I had misunderstood the API a little. One isn't supposed to
fill the buffer by pointing to new data on every cycle. One is
supposed to do so only after getting the need_buffer error, which
means we need to alternate between sending data and getting an error.

Adapt to this, and also change the Trace messages in the same commit
to log read/written data even if we get errors.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>