feat(artifact/payload): Support multiple files read in each payload
This adds support for reading multiple files from each payload.

The solution is just an extension of the existing `Next()` interface, so that
now:

```
artifact::Parse() -> artifact

artifact.Next() -> Payload

Payload.Next() -> payload::Reader

...

Payload.Next() -> NoMorePayloadFilesError
```

Hence, each payload, and each file in the given payload can be processed in
turn.

The reader verifies the checksum of the given file as it is consumed, and
returns an error if there is a checksum mismatch.

The `payload::Reader` also has accessors for getting the `filename` and the
`size` of the given file.

Ticket: MEN-6452
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>