refac(artifact): Add the Parse interface to the toplevel artifact::
`parser::` should not be exposed outside of the artifact:: namespace. Hence we
alias the `parser::Artifact`, and wrap the `Next` method, to provide the
top-level interface in `artifact`.

All elements of the payload:: namespace is also exposed to the end user as a
part of the interface.

This includes the `Payload` reader, and the payload file reader, as well as it's
expected equivalents.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>