feat(http): Add https support to the client
This adds support for HTTPs to the http library.

SSL is optionally enabled based on the protocol of the URL given to the
`AsyncCall` function.

The client uses the system ca-cert store by default, and if a certificate is
given to the client config, then this is also added to the certification pool.

Ticket: MEN-6144, MEN-6124
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>