chore: Improve processes::Wait() and exit status fetching.
Turn `Wait()` into returning an error. This is mainly in order to make
it work better inside functions which themselves return errors.
Otherwise it constantly has to be wrapped in code which produces an
error out of it. The status code can still be fetched explicitly via
`GetExitStatus()`.

Also beef up the tests a little bit, particularly around process
killing.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>