fix: Fix C++17 flag being mistakely propagated to many cross-platform parts.
The reason is the `INTERFACE` option which is removed in this commit:
It causes all users of the library to use the flag, even
cross-platform code. In fact, we don't need C++17 at all in the
relevant source files so just remove it altogether (otherwise the
correct fix would have been `PRIVATE`).

This triggered a few errors since we had been compiling with the wrong
flag for a while.

* Fix binary incompatibility in optional-lite (see code comment).

* Remove member initializer from `ParserConfig`. The reason is that
  aggregate initialization and member initialization conflict in
  C++11, but not in C++14 and later (see [1]). Since the
  initialization was empty anyway, the behavior is the same even if we
  remove it.

[1] https://en.wikipedia.org/wiki/C%2B%2B14#Aggregate_member_initialization

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>