chore: Add `ExceptionToErrorOrAbort`.
Some parts of the C++ API depend on exceptions. If they are turned
off then the program will just abort. This is going to be used for
regex in particular, where we can control the outcome, so it should
not normally happen at all. But with this function we can at least
avoid catastrophic failures on platforms where they are supported.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>