fix: Fix outside-stack memory read error due to default move constructor.
This one was sneaky. StringReader was being move constructed using the
default constructor. Which is fine, both members support move
construction as well. However, one important detail is that one member
takes the other member as an argument. Which means that after they are
moved, although both are valid objects, the reference inside reader is
now pointing to the member of the old StringReader.

StreamReader doesn't provide an interface for changing the reference
after it has been constructed (and it's probably best to keep it this
way), so fix it by creating a new StreamReader instead.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>