fix: Add missing handling for destroyed client for `async_resolve`.
This was simply an oversight. All the other async handlers are testing
`weak_client` in order to avoid calling handlers for Clients that have
been destroyed. This was missing for `async_resolve`, leading to
crashes if the client was destroyed before the resolve phase was over.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>