chore(crypto): harvest the OpenSSL error message for all failed calls
Since the OpenSSL errors are stored in a queue, and harvested in a FIFO like
fashion, it is important that we do harvest the messages upon every call which
fails, otherwise the extraction will generate the wrong error messages.

Therefore we loop over, and extract all errors from the queue on each call to
`GetOpenSSLErrorMessage`.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>