fix: Fix incorrect reuse of stream and parser.
When reusing `http::Client` for multiple requests, we don't keep track
of what state the stream and the parser were in before starting over.
This would sometimes, but not always, lead to errors about incomplete
requests. So we need to recreate those with fresh instances to be
safe.

More efficient implementations would take the state into account and
reuse connections when possible, but this is an optimization for the
future.

This required some other changes too. We need the `ssl_context` every
time we create a stream, so move that from `ClientConfig` into
`Client`. This is where it probably should be anyway, since
`ClientConfig` should not expose any Boost structures.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>