fix: Fix race condition in AsyncWait for quick processes.
If the process finishes so quickly that `SetupAsyncWait` completes
before the caller has a chance to call `AsyncWait`, then the handler
is never called. With this commit, this scenario is detected and the
handler is scheduled right away if we find that the process has
already ended.

This showed up on my system occasionally as a very rare hang in the
Update Module state calls tests.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>