fix: Don't install `version` file, it should be installed by users.
The version is not there to describe which version of Mender you have.
It's there to describe which version of state scripts you have. And
then the client must check that it supports that version. So the file
should be installed by the user or the build system which handles
state scripts.

In Yocto it is installed as part of custom state scripts.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>