chore: Revert "build(artifact-scripts): Add the Artifact script version file to install"
This reverts commit 8be204fa7696da2cf079667de12775aab87822ff.

The version is not there to describe which version of Mender you have.
It's there to describe which version of state scripts you have. And
then the client must check that it supports that version. So the file
should be installed by the user or the build system which handles
state scripts.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>