chore: Extend `crypto` library to support RSA signature verification
The library is extended with the new method `VerifySign` that verifies a
signature given the public key and the digest of the message.

To keep the interface consistent, the existing function `Sign` which took
the raw message as an input is moved to `SignRawData` and a new `Sign`
function is implemented for signing pre-computed digests. Pass also
const reference in these functions where possible.

Additionally:
* Functions to encode and decode Base64 are implemented/exposes
* Docstring is added to `EncodeBase64` to clarify the length prediction
* The `mender::sha::SHA` class now knows the length of the digest

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>