feat: Artifact manifest RSA signature parse and verification
Implement artifact/v3/manifest_sig to parse and hold the optional
signature of an Artifact. It is verified by the main Artifact parser at
`Parse` time if public keys were provided by the configuration.

The `Manifest` class has been modified to parse and hold a SHA256 digest
of the contents so that the signature can be verified later on.

Ticket: MEN-6203
Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>