chore: Simplify AsyncReaderFromReader and get rid of extra thread.
Same for AsyncWriterFromWriter.

Threads are unlikely to be allowed on low level embedded devices, and
we don't strongly need it anyway, if we just accept that certain waits
are synchronous. There is still time to process events in between each
returned read, so it's not too bad. And we can always improve it later
since the interface is the same.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>