docs: Tweak and clarify some things in `SupportsRollback` query.
Let's take this major version opportunity to change a very small and
insignificant detail in the Update Module specification: When
`SupportsRollback` is called. This does not really matter in practice,
so let's relax this condition.

Also remove a response that was never implemented, and which is
unlikely to ever be (even more now that rootfs updates are modules).

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>