chore: Set maximum state transitions to a reasonable number.
There is nothing accurate about this number. It's just a number which
won't loop too many times, but still gives some wiggle room to add a
few more states without exceeding the number. It was determined simply
by matching the number of loops in one of the tests which tries to
exceed it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>