chore: Change HTTP API to be more subclass friendly.
First of all, change the main interface to use only `AsyncReaders` for
bodies. Move this interface to the main `Client` and `Server` classes,
as opposed to the existing request and response classes, so that it's
easier to subclass and modify the behavior. We still provide the API
directly on the request and response objects for convenience, but
these just call the main API, so they are still overridable.

Also, keep the `SetBodyWriter` interface on the request and response
objects, since it is used a lot. A `Writer` is no longer explicitly
supported by the HTTP API, but instead we implement it as a
convenience function to automatically and asynchronously copy data
from the reader to the writer.

As part of this, consuming the body is now no longer optional, since
the `AsyncReader` has to drive the download process. The resulted in a
lot of tests, that don't need the body, writing the body to
`io::Discard`.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>