chore(http): Standardize cancellation handling on one variable.
Previously we had the `client_active_` and the `stream_active_`
variables, but these had overlapping roles with `cancelled_`, which is
to detect whether the request/response is still active. Instead of
that confusing double-accounting, switch to rely on `cancelled_` only.

This solves some complex destruction scenarios where cancellers would
call destructors, which called back into the original and such.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>