fix: Ignore sender when matching signal
The sender set on a received DBus message is the unique name (e.g
":1.0") not the well-known name the sender requested and
acquired. Since the unique name is not predictable, we cannot
match it.

Also define a static inline function for getting a signal
spec/match rule to avoid code repetition.

Ticket: MEN-6651
Changelog: none
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>