test: Skip DBus tests when built with -fsanitize=thread
The thread sanitizer doesn't like what libdbus is doing with
two of its locks. In particular, it thinks that locking and
unlocking M1 while holding a lock on M0 and then, much later,
locking and unlocking M0 while holding a lock on M1 is a
problem. So a sequence like this:

```
1. lock M0
2. lock M1
3. unlock M1
4. unlock M0
...
w. lock M1
x. lock M0
y. unlock M0
z. unlock M1
```

In our case, all this happens in a single thread so it's 100%
safe. But the thread sanitizer still sees it as problem and since
it is not in our code, there doesn't seem to be a working way to
suppress the warning. Hence, skip these tests in ThreadSan
builds/runs, we declare our DBus API/implementation thread-unsafe
anyway.

Ticket: MEN-6653
Changelog: none
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>