refac(config): Collect all configuration in the conf package
This splits out the http-client configuration from the `client` package, and
moves it into the `conf` package, to remove the dependency on the whole `client`
package, when using the `conf` package.

This was forced upon me during the process of splitting out `mender-setup` and
`mender-snapshot` into each their own separate binaries.

Ticket: MEN-6542

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>