fix(state-machine): Go right to Idle state after a deployment
Previously, the flow would do a `Sync_Leave` after finishing an update, which is
not right.

The fix is simple - just go to `idle_enter` right after an update.

Also, do still send the inventory right after an update, but do it through
adding an event to trigger the regular state machinery to handle it.

Ticket: MEN-6671
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>