refac(crypto): Change SignData to accept all OpenSSL URLs
This is done through using the already existing `PrivateKey` abstraction, as
opposed to using `BIO` `OpenSSL` functions directly.

The change should be transparent to the already existing functionality, as
keyfiles are now just a subset of the way to store and retrieve files, and we
don't really care how the `private-keys` are stored.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>