refac: Refactor the whole `common::cli` inside `common::conf`
The first had some problems, like the interdependency between
`common::cli` and `common:conf` and the horrible code duplication
between `mender::upate::cli` and `mender::auth::cli`.

From a different point of view, the parsing of the (common) CLI options
is done in `common::conf` for both tools, which suggest that there is
reason to have everything in `common::conf` in the first place.

While merging the global options, add the missing `--version` option and
do minor beautifications.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>