fix: Use our custom false-initialized boolean in http::ClientConfig
Turns out `bool` values are not guaranteed to be initialized to
`false` by default. However, if we explicitly initialize fields
in `http::ClientConfig`, we cannot use designated (named)
initializers. So we need a custom boolean type that acts like a
boolean, but is guaranteed to be initialized to `false` by
default.

Ticket: MEN-6787
Changelog: none
Signed-off-by: Vratislav Podzimek <v.podzimek@mykolab.com>