feat(watchdog): Add application watchdog support for the systemd service
This adds support for an application watchdog through the systemd's supported
`WatchdogSec=` feature. In which the application will have to `ping` systemd, or
the service will restart.

Ticket: MEN-6763
Changelog: Add application watchdog support for the systemd service by default.
The default value is set to 24-hours, and if `mender-update` has not given any
sign of life within this interval, the service will be restarted by `systemd`.
The timeout value can be configured through `WatchdogSec=` in the
`mender-updated` service file, removed, or the feature can be disabled through
`MENDER_USE_SYSTEMD` when building. It can also be configured through the Yocto
build, if you desire another interval set.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
Signed-off-by: Ole P. Orhagen <ole.orhagen@northern.tech>