chore: Immediately give up reader/writer refs after AsyncCopy is complete.
I'm not sure about the exact reason, but the Boost event loop seems to
keep handlers alive somewhere a bit longer than we expect, but it's
important for us that once `AsyncCopy` is finished, it does not keep
its references, so that things get destroyed at the time we expect
them to. So enforce this when the handler is called.

This requires us to change a few tests which were (somewhat
unintentionally) relying on the lifetime of the readers/writers to
extend beyond the finishing of `AsyncCopy`. It also means that the
error mode argument to `SetBodyWriter` is not appropriate anymore. The
connection will always be cancelled because the reader is destroyed
afterwards.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>