chore(state-machine): Panic in the case of non-handled non-deferred events
This adds a sanity check to verify that there are no non-deferred events left in
the event queue after the run_queue has been populated.

If there was, then we would have undefined behaviour in the state-machine, and
as such the best thing we can do in this instance is to panic. This is a serious
logic error in our code, and as such such cause a hard-fault.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>