chore(mender-updated): Remove the Conflicts with mender-client
The service file `Conflicts` works a little different than expected, from the
man page:

```
Conflicts=
A space-separated list of unit names. Configures negative requirement dependencies. If a
unit has a Conflicts= setting on another unit, starting the former will stop the latter
and vice versa.
```

This did not work so well with the newly created alias to `mender-client` in the
`mender-authd` service file.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>