test: Verify that identity is marshaled the same way as in Go
This adds a test to verify that identity is marshaled in the same way as it is
in the Golang client code.

See the test here:
`https://github.com/mendersoftware/mender/blob/master/device/identity_data_test.go#L24`

Note that we still might have a UTF-8 issue here. But this is a separate topic.

Ticket: MEN-6825

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>