feat(state-scripts): Support running version 2 scripts
There are no differences between the version `2`, and the version `3` of the
state scripts. The bump was done, simply to keep them in line with the version
of the Artifact format.

Supporting version `2` is necessary, if we are to support upgrade paths from
`1.x` line clients.

Ticket: None
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>