chore: Make sure tar reader always exhausts its archive source.
Logically, we don't need to do this, because the tar reader already
knows that it's at the end, so we don't need additional confirmation.
However, there is a reason why we still want to do it: If the HTTP
body reader does not get a final read, then it complains that the
reader is destroyed before the stream has been fully consumed. This is
quite an annoying warning, especially in standalone mode.

Regarding the changed case labels: The `archive_read_data` call
actually cannot return the `ARCHIVE_OK` and `ARCHIVE_EOF` codes, only
the error codes. Instead it returns number of bytes read, and 0 for
EOF (`ARCHIVE_OK` happens to also be zero, so this was a source of
confusion).

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>