chore: Make ReaderFromAsyncReader safe to use with immediate completions.
If `AsyncRead` happened to call the handler immediately
(synchronously, and before returning), because the result is already
known, then the operation would hang. This happened because we would
then stop the loop before starting it. Although intuitively this still
balances correctly, the event loop does not support stopping first,
and then running, it must be run first, then stopped.

Fix this by running the inner `AsyncRead` in an event loop handler
instead, so that the loop is already running by the time we get there.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>