chore: Fix several issues in artifact parser.
At the moment we only support one payload. This is alright, but it
means that the parser makes some assumptions which are not ideal. To
start with, it reads the next payload when the artifact parsing
starts, but it should not read it until we call `Next()`. This isn't
required for simply reading one single payload, but is required if we
want to use `Next()` again to verify that there indeed are no more
payloads, which brings us to the next issue.

It's generally good practice to make sure there is no garbage at the
end of the artifact stream, including additional payloads. While it's
unlikely to be exploitable, it's always good to be sure. But this is
not the main reason, the main reason is that if we don't read all the
way to the end, including the EOF, then HTTP complains that the stream
is closed before it has been fully consumed. This is a message we'd
like to retain for other reasons, and this message would be triggered
on every single install, so it looks quite bad.

Now both the daemon and standalone modes call `artifact.Next()` until
it is exhausted.

And finally, don't mistake an Unrecognized artifact for an EOF token.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>